app/requirements.txt
Flask==3.0.3
paramiko
watchdog
livereload
pymongo
ansible
ansible-runner
cryptography
python-dotenv